% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra}
\usage{
dijkstra(graph, start, end)
}
\arguments{
\item{graph}{matrix containing the graph}

\item{start}{id of the start point}

\item{end}{id of the end point}
}
\value{
a list of edges that are part of the shortest path
}
\description{
Calculates the shortest path between two points on a graph using Dijkstra's
algorithm
}
