% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Calculate a distance matrix between all pairs of a given list of points}
\usage{
distance_matrix(graph, xy)
}
\arguments{
\item{graph}{Graphs extracted from \link{download_graph}.}

\item{xy}{Matrix of two columns containing latitudes and longitudes of points
between which distances are to be calcualted.}
}
\value{
A list of two items: A matrix of distances between all pairs of
points listed in \code{xy}, and an index of points in \code{xy} which map on
to unique graph nodes.
}
\description{
Calculate a distance matrix between all pairs of a given list of points
}
\note{
Different points may map on to the same network locations, in which
case they are excluded from distance calculation. The function returns an
index which can be used to directly extract the \code{xy} points which map on
to unique locations.
}
