% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-graph.R
\name{download_graph}
\alias{download_graph}
\title{Download OSM road graph and preprocess it}
\usage{
download_graph(start_pt, end_pt, weighting_profile = "bicycle", buffer = 0)
}
\arguments{
\item{start_pt}{Two numeric values (latitude, longitude) as start point
coordinates.}

\item{end_pt}{Two numeric values (latitude, longitude) as end point
coordinates.}

\item{weighting_profile}{Name of the used weighting profile.
\code{osmprob::weighting_profiles} contains all available profiles.}

\item{buffer}{Positive value that defines by how much (in percent) should the
downloaded data extend the bounding box defined by \code{start_pt} and
\code{end_pt}.}
}
\value{
graphs \code{list} containing the original street graph, a minimized
graph map linking the two to each other.
}
\description{
This function uses \code{osmdata} to download a OSM street graph. The extent
of the downloaded area is defined by the start and end coordinates of the
desired path plus a buffer, that extends the bounding box in all directions.
The graph is then preprocessed for routing purposes, which includes removing
topologically unnecessary parts and calculating edge weights based on the
selected weighting profile.
}
\examples{
\dontrun{
start_pt <- (11.580, 48.140)
end_pt <- (11.585, 48.145)
graph <- download_graph (start_pt = start_pt, end_pt = end_pt,
weighting_profile = "bicycle", buffer = 0)
}
}
