% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{get_probability}
\alias{get_probability}
\title{Calculate routing probabilities for a data.frame}
\usage{
get_probability(graphs, start_node, end_node, eta = 1)
}
\arguments{
\item{graphs}{\code{list} containing the two graphs and a map linking the two
to each other.}

\item{start_node}{Starting node for shortest path route}

\item{end_node}{Ending node for shortest path route}

\item{eta}{The parameter controlling the entropy (scale is arbitrary)}
}
\description{
Calculate routing probabilities for a data.frame
}
\examples{
\dontrun{
  dat <- readRDS ("../compact-ways-munich.Rda") \%>\% make_compact_graph
  st <- dat$compact$from_id [1]
  en <- dat$compact$to_id [10]
  get_probability (dat, st, en)
}
}
