% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Download osm data and calculate street lengths}
\usage{
getData(bbox = c(-0.27, 51.47, -0.2, 51.5))
}
\arguments{
\item{bbox}{Numeric bounding box}
}
\value{
data.frame containing the osm data
}
\description{
Download osm data and calculate street lengths
}
