% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDist}
\alias{getDist}
\title{getDist}
\usage{
getDist(xAll, yAll)
}
\arguments{
\item{xAll}{vector containing all latitudes sequentially}

\item{yAll}{vector containing all longitudes sequentially}
}
\value{
the total distance between a and b in km
}
\description{
Calculates the total distance in km between all sequential points in xAll
and yAll
}
