% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router-mp.R
\name{getProbability}
\alias{getProbability}
\title{Calculate routing probabilities for a data.frame}
\usage{
getProbability(netdf, start_node, end_node, eta = 1)
}
\arguments{
\item{netdf}{\code{data.frame} of network}

\item{start_node}{Starting node for shortest path route}

\item{end_node}{Ending node for shortest path route}

\item{eta}{The parameter controlling the entropy (scale is arbitrary)}
}
\description{
Calculate routing probabilities for a data.frame
}
