% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{make_compact_graph}
\alias{make_compact_graph}
\title{Removes nodes and edges from a graph that are not needed for routing}
\usage{
make_compact_graph(graph)
}
\arguments{
\item{graph}{\code{data.frame} of the graph to be processed.}
}
\value{
\code{data.frame} containing the output graph.
}
\description{
\code{make_compact_graph} takes a \code{data.frame} containing a graph and
removes all vertices not needed for routing as well as all vertices that are
not connected to the largest coherent part of the graph.
}
\examples{
\dontrun{
q <- osmdata::opq (bbox = c (11.58, 48.14, 11.585, 48.145))
q <- osmdata::add_feature (q, key = 'highway')
x <- osmdata::osmdata_sf(q)
netw <- osmlines_as_network (x)
graph <- make_compact_graph (graph = netw)
}

}
