% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router-mp.R
\name{osm_router}
\alias{osm_router}
\title{router test}
\usage{
osm_router(netdf, start_node, end_node, eta = 1)
}
\arguments{
\item{netdf}{\code{data.frame} of network}

\item{start_node}{Starting node for shortest path route}

\item{end_node}{Ending node for shortest path route}

\item{eta}{The parameter controlling the entropy (scale is arbitrary)}
}
\description{
router test
}
\note{
Function called 'osm_router' to avoid ambiguities with all other things
that might be called 'router'. \code{netdt} must have just three columns of
\code{xfr}, \code{xto}, and \code{d}, where the first two are simple integer
indices.
}
\examples{
netdf <- data.frame (
                     xfr = c (rep (0, 3), rep (1, 3), rep (2, 4),
                              rep (3, 3), rep (4, 2), rep (5, 3)),
                     xto = c (1, 2, 5, 0, 2, 3, 0, 1, 3, 5,
                              1, 2, 4, 3, 5, 0, 2, 4),
                     d = c (7., 9., 14., 7., 10., 15., 9., 10., 11., 2.,
                            15., 11., 6., 6., 9., 14., 2., 9.))
osm_router (netdf, 1, 5)
}
