% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmlines-as-network.R
\name{osmlines_as_network}
\alias{osmlines_as_network}
\title{Convert osm_lines to a data.frame of sequential network connections}
\usage{
osmlines_as_network(lns, profile_name = "bicycle")
}
\arguments{
\item{lns}{An \code{sf} collection of \code{LINESTRING} objects, obtained for
example from the \code{osm_lines} component of an \code{osmdata} object}

\item{profile_name}{Name of the used weighting profile.
\code{osmprob::weighting_profiles} contains all available profiles.}
}
\value{
\code{data.frame} of all pairs of connected nodes
}
\description{
Convert osm_lines to a data.frame of sequential network connections
}
\examples{
\dontrun{
q <- osmdata::opq (bbox = c (11.58, 48.14, 11.585, 48.145))
q <- osmdata::add_feature (q, key = 'highway')
x <- osmdata::osmdata_sf(q)
net <- osmlines_as_network (x)
}
}
