% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmprob-package.R
\docType{package}
\name{osmprob}
\alias{osmprob}
\alias{osmprob-package}
\title{Description: Calculates probabilistic routes on a OSM street graph}
\description{
Downloads and preprocesses OpenStreetMap (OSM) data using \code{osmdata} and
performs probabilistic routing on the graph. The result is a
\code{data.frame}, where each row represents an edge of the graph that has
one column containing the local and one column containing the global
traversal probability. This data can either be used for further analysis or
displayed in a web broser using a built-in \code{shiny}/\code{leaflet}
function.
}
