% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-graph.R
\name{plot_map}
\alias{plot_map}
\title{Plots the graph network as a Shiny Leaflet app in a browser.}
\usage{
plot_map(graph, shortest)
}
\arguments{
\item{graph}{\code{data.frame} containing the street graph to be
displayed.}

\item{shortest}{\code{vector} containing the shortest path.}
}
\description{
Plots the graph network as a Shiny Leaflet app in a browser.
}
\examples{
\dontrun{
q <- osmdata::opq (bbox = c (11.58, 48.14, 11.585, 48.145))
q <- osmdata::add_feature (q, key = 'highway')
x <- osmdata::osmdata_sf(q)
net <- osmlines_as_network (x)
graph <- make_compact_graph (net)
start_pt <- graph$from_id [1]
end_pt <- graph$to_id [100]
prob <- get_probability (graph, start_pt, end_pt)
short <- get_shortest_path (graph, start_pt, end_pt)
plot_map (prob, short)
}

}
