% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_makeCompactGraph}
\alias{rcpp_makeCompactGraph}
\title{rcpp_makeCompactGraph}
\usage{
rcpp_makeCompactGraph(graph)
}
\arguments{
\item{graph}{graph to be processed}
}
\value{
Rcpp::DataFrame containing the output graph
}
\description{
Removes nodes and edges from a graph that are not needed for routing
}
