% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_router}
\alias{rcpp_router}
\title{rcpp_router}
\usage{
rcpp_router(netdf, start_node, end_node)
}
\arguments{
\item{netdf}{\code{data.frame} containing network connections}

\item{start_node}{Starting node for shortest path route}

\item{end_node}{Ending node for shortest path route}
}
\value{
Rcpp::List objects of OSM data
}
\description{
Return OSM data in Simple Features format
}
