% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{road_data_sample}
\alias{road_data_sample}
\title{Sample road data}
\format{\code{list} with three \code{data.frame}. One containing the original
graph, one the compact graph and one the pairs of edge IDs needed to map the
two graphs to one another.}
\usage{
road_data_sample
}
\description{
Sample data set of road data in Munich, Germany that has been downloaded and
preprocessed, so it can be used for routing.
}
\references{
\url{http://www.openstreetmap.org/}
}
\keyword{datasets}
