% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_vertices_by_coordinates}
\alias{select_vertices_by_coordinates}
\title{Select vertices on graph that are closest to the specified coordinates.}
\usage{
select_vertices_by_coordinates(graph, start_coords, end_coords)
}
\arguments{
\item{graph}{\code{data.frame} containing the street network.}

\item{start_coords}{\code{numeric} coordinates of the start point.}

\item{end_coords}{\code{numeric} coordinates of the end point.}
}
\value{
\code{list} containing the two rows of the input \code{data.frame}
that are closest to the start and end coordinates
}
\description{
Select vertices on graph that are closest to the specified coordinates.
}
