% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weightingProfiles}
\alias{weightingProfiles}
\title{Weighting profiles}
\format{A rda file}
\usage{
weightingProfiles
}
\description{
Collection of weighting profiles used in the routing process. Original data
taken from the Routino project.
}
\references{
\url{https://www.routino.org/xml/routino-profiles.xml}
}
\keyword{datasets}
