% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{get_probability}
\alias{get_probability}
\title{Calculate routing probabilities for a data.frame}
\usage{
get_probability(graphs, start_node, end_node, eta = 1)
}
\arguments{
\item{graphs}{\code{list} containing the two graphs and a map linking the two
to each other.}

\item{start_node}{Starting node for shortest path route}

\item{end_node}{Ending node for shortest path route}

\item{eta}{The parameter controlling the entropy (scale is arbitrary)}
}
\value{
\code{list} containing the \code{data.frame} of the graph elements
with the routing probabilities and the estimated probabilistic distance.
}
\description{
Calculate routing probabilities for a data.frame
}
\examples{
\dontrun{
  graph <- road_data_sample
  start_pt <- c (11.603,48.163)
  end_pt <- c (11.608,48.167)
  pts <- select_vertices_by_coordinates (graph, start_pt, end_pt)
  route_start <- pts[1]
  route_end <- pts [2]
  get_probability (graphs = graph, start_node = route_start,
  end_node = route_end, eta = 0.6)
}
}
