% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{get_shortest_path}
\alias{get_shortest_path}
\title{Calculate the shortest path between two nodes on a graph}
\usage{
get_shortest_path(graphs, start_node, end_node)
}
\arguments{
\item{graphs}{\code{list} containing the two graphs and a map linking the two
to each other.}

\item{start_node}{Starting node for shortest path route.}

\item{end_node}{Ending node for shortest path route.}
}
\value{
\code{data.frame} of the graph elements the shortest path lies on.
}
\description{
Calculate the shortest path between two nodes on a graph
}
\examples{
\dontrun{
  dat <- readRDS ("../compact-ways-munich.Rda") \%>\% make_compact_graph
  st <- dat$compact$from_id [1]
  en <- dat$compact$to_id [10]
  get_shortest_path (graphs = dat, start_node = st, end_node = en)
}
}
