% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router-mp.R
\name{getProbability}
\alias{getProbability}
\title{Calculate routing probabilities for a data.frame}
\usage{
getProbability(netdf, start_node, end_node, eta = 1)
}
\arguments{
\item{netdf}{\code{data.frame} of network}

\item{start_node}{Starting node for shortest path route}

\item{end_node}{Ending node for shortest path route}

\item{eta}{The parameter controlling the entropy (scale is arbitrary)}
}
\description{
Calculate routing probabilities for a data.frame
}
\examples{
netdf <- data.frame (
                     from_id = c (rep (0, 3), rep (1, 3), rep (2, 4),
                              rep (3, 3), rep (4, 2), rep (5, 3)),
                     to_id = c (1, 2, 5, 0, 2, 3, 0, 1, 3, 5,
                              1, 2, 4, 3, 5, 0, 2, 4),
                     d_weighted = c (7., 9., 14., 7., 10., 15., 9., 10.,11.,
                              2., 15., 11., 6., 6., 9., 14., 2., 9.))
prb <- getProbability (netdf, 1, 5)
\dontrun{
library (magrittr)
devtools::load_all (export_all = TRUE)
netdf <- readRDS ("../tests/compact-ways-munich.Rda") \%>\% head (115) \%>\% makeCompactGraph
netdf$d_weighted <- netdf$d
start_node <- netdf$from_id [20]
end_node <- netdf$to_id [50]
eta <- 1
}
}
