% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCompactGraph.R
\name{makeCompactGraph}
\alias{makeCompactGraph}
\title{Removes nodes and edges from a graph that are not needed for routing}
\usage{
makeCompactGraph(graph)
}
\arguments{
\item{graph}{\code{data.frame} of the graph to be processed}
}
\value{
\code{data.frame} containing the output graph
}
\description{
\code{makeCompactGraph} takes a \code{data.frame} containing a graph and
removes all vertices not needed for routing as well as all vertices that are
not connected to the largest coherent part of the graph.
}
