% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrimmedGraph.R
\name{plotMap}
\alias{plotMap}
\title{Plot the graph network as a Shiny Leaflet app in a browser.}
\usage{
plotMap(graph, shortest)
}
\arguments{
\item{graph}{\code{data.frame} containing the street graph to be
displayed}

\item{shortest}{\code{vector} containing the shortest path}
}
\description{
Plot the graph network as a Shiny Leaflet app in a browser.
}
